Ignore more pylint checks in lp3 light mode.
Ignore additional pylint checks in lp3's when passing `-l|--light`.

Also refactor lp3 to have an array of pylint error codes to ignore,
similar to the selection of checks for flake8.

Also move the array definitions for light mode into the if-clause in
order to only define the variables when necessary.

Also add shellcheck ignores to avoid shellcheck complains on lines that
purposefully don't quote variables.